@@ -59,13 +59,13 @@ extensions = [
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3.9", None),
     "numpy": ("https://numpy.org/doc/stable", None),
-    "scipy": ("https://docs.scipy.org/doc/scipy/reference", None),
+    "scipy": ("https://docs.scipy.org/doc/scipy", None),
     "scikit.learn": ("https://scikit-learn.org/stable", None),
     "openmm": ("http://docs.openmm.org/latest/api-python/", None),
     "rdkit": ("https://www.rdkit.org/docs", None),
     "openeye": ("https://docs.eyesopen.com/toolkits/python/", None),
     "mdtraj": ("https://www.mdtraj.org/1.9.5/", None),
-    "openff.units": ("https://docs.openforcefield.org/units/en/stable", None),
+    "openff.units": ("https://docs.openforcefield.org/projects/units/en/stable", None),
     "gufe": ("https://gufe.readthedocs.io/en/latest/", None),
 }
 
@@ -186,17 +186,19 @@ except Exception as e:
         f"Getting ExampleNotebooks failed in {filename} line {lineno}: {e}"
     )
 
+
+# First, create links at top of notebook pages
+# All notebooks are in ExampleNotebooks repo, so link to that
+# Finally, add sphinx reference anchor in prolog so that we can make refs
 nbsphinx_prolog = cleandoc(r"""
-    {%- set path = env.doc2path(env.docname, base="ExampleNotebooks") -%}
-    {%- set gh_repo = "OpenFreeEnergy/openfe" -%}
+    {%- set gh_repo = "OpenFreeEnergy/ExampleNotebooks" -%}
     {%- set gh_branch = "main" -%}
     {%- set path = env.doc2path(env.docname, base=None) -%}
     {%- if path.endswith(".nblink") -%}
         {%- set path = env.metadata[env.docname]["nbsphinx-link-target"] -%}
     {%- endif -%}
-    {%- if path.startswith("docs/ExampleNotebooks") -%}
-        {%- set path = path.replace("docs/ExampleNotebooks/", "", 1) -%}
-        {%- set gh_repo = "OpenFreeEnergy/ExampleNotebooks" -%}
+    {%- if path.startswith("ExampleNotebooks/") -%}
+        {%- set path = path.replace("ExampleNotebooks/", "", 1) -%}
     {%- endif -%}
     {%- set gh_url =
         "https://www.github.com/"
@@ -246,4 +248,5 @@ nbsphinx_prolog = cleandoc(r"""
 
             :octicon:`rocket` Run in Colab
 
+    .. _{{ env.doc2path(env.docname, base=None) }}:
 """)
