@@ -11,7 +11,8 @@ from __future__ import annotations
 
 from typing import Optional, Literal
 from openff.units import unit
-from openff.models.types import FloatQuantity
+from openff.models.types import FloatQuantity, ArrayQuantity
+from openff.interchange.components._packmol import _box_vectors_are_in_reduced_form
 
 from gufe.settings import (
     Settings,
@@ -21,46 +22,188 @@ from gufe.settings import (
 )
 
 
-try:
-    from pydantic.v1 import validator
-except ImportError:
-    from pydantic import validator  # type: ignore[assignment]
+from pydantic.v1 import validator
 
 
 class BaseSolvationSettings(SettingsBaseModel):
     """
-    Base class for SolvationSettings objects
+    Base class for SolvationSettings objects.
     """
     class Config:
         arbitrary_types_allowed = True
 
 
 class OpenMMSolvationSettings(BaseSolvationSettings):
-    """Settings for controlling how a system is solvated using OpenMM tooling
+    """Settings for controlling how a system is solvated using OpenMM tooling.
 
-    Note
-    ----
-    No solvation will happen if a SolventComponent is not passed.
+    Defining the number of waters
+    -----------------------------
 
+    The number of waters is controlled by either:
+      a) defining a solvent padding (``solvent_padding``) in combination
+         with a box shape
+      b) defining the number of solvent molecules
+         (``number_of_solvent_molecules``)
+         alongside the box shape (``box_shape``)
+      c) defining the box directly either through the box vectors
+         (``box_vectors``) or rectangular box lengths (``box_size``)
+
+    When using ``solvent_padding``, ``box_vectors``, or ``box_size``,
+    the exact number of waters added is determined automatically by OpenMM
+    through :meth:`openmm.app.Modeller.addSolvent` internal heuristics.
+    Briefly, the necessary volume required by a single water is estimated
+    and then the defined target cell is packed with waters avoiding clashes
+    with existing solutes and box edges.
+
+
+    Defining the periodic cell size
+    -------------------------------
+
+    The periodic cell size is defined by one, and only one, of the following:
+      * ``solvent_padding`` in combination with ``box_shape``,
+      * ``number_of_solvent_molecules`` in combination with ``box_shape``,
+      * ``box_vectors``,
+      * ``box_size``
+
+    When using ``number_of_solvent_molecules``, the size of the cell is
+    defined by :meth:`openmm.app.Modeller.addSolvent` internal heuristics,
+    automatically selecting a padding value that is large enough to contain
+    the number of waters based on a geometric estimate of the volume required
+    by each water molecule.
+
+
+    Defining the periodic cell shape
+    ---------------------------------
+
+    The periodic cell shape is defined by one, and only one, of the following:
+      * ``box_shape``,
+      * ``box_vectors``,
+      * ``box_size``
+
+    Default settings will create a cubic box, although more space efficient
+    shapes (e.g. ``dodecahedrons``) are recommended to improve simulation
+    performance.
+
+
+    Notes
+    -----
+    * The number of water molecules added will be affected by the number of
+      ions defined in SolventComponent. For example, the value of
+      ``number_of_solvent_molecules`` is the sum of the number of counterions
+      added and the number of water molecules added.
+    * Solvent addition does not account for any pre-existing waters explicitly
+      defined in the :class:`openfe.ChemicalSystem`. Any waters will be added
+      in addition to those pre-existing waters.
+    * No solvation will happen if a SolventComponent is not passed.
+
+
+    See Also
+    --------
+    :mod:`openmm.app.Modeller`
+    Base class for SolvationSettings objects
     """
     solvent_model: Literal['tip3p', 'spce', 'tip4pew', 'tip5p'] = 'tip3p'
     """
-    Force field water model to use.
+    Force field water model to use when solvating and defining the model
+    properties (e.g. adding virtual site particles).
+
     Allowed values are; `tip3p`, `spce`, `tip4pew`, and `tip5p`.
     """
+    solvent_padding: Optional[FloatQuantity['nanometer']] = 1.2 * unit.nanometer
+    """
+    Minimum distance from any solute bounding sphere to the edge of the box.
 
-    solvent_padding: FloatQuantity['nanometer'] = 1.2 * unit.nanometer
-    """Minimum distance from any solute atoms to the solvent box edge."""
+    Note
+    ----
+    * Cannot be defined alongside ``number_of_solvent_molecules``,
+      ``box_size``, or ``box_vectors``.
+    """
+    box_shape: Optional[Literal['cube', 'dodecahedron', 'octahedron']] = 'cube'
+    """
+    The shape of the periodic box to create.
+
+    Notes
+    -----
+    * Must be one of `cube`, `dodecahedron`, or `octahedron`.
+    * Cannot be defined alongside ``box_vectors`` or ``box_size``.
+    """
+    number_of_solvent_molecules: Optional[int] = None
+    """
+    The number of solvent molecules (water + ions) to add.
+
+    Note
+    ----
+    * Cannot be defined alongside ``solvent_padding``, ``box_size``,
+      or ``box_vectors``.
+    """
+    box_vectors: Optional[ArrayQuantity['nanometer']] = None
+    """
+    `OpenMM reduced form box vectors <http://docs.openmm.org/latest/userguide/theory/05_other_features.html#periodic-boundary-conditions>`.
+
+    Notes
+    -----
+    * Cannot be defined alongside ``solvent_padding``,
+      ``number_of_solvent_molecules``, or ``box_size``.
+
+    See Also
+    --------
+    :mod:`openff.interchange.components.interchange`
+    :mod:`openff.interchange.components._packmol`
+    """
+    box_size: Optional[ArrayQuantity['nanometer']] = None
+    """
+    X, Y, and Z lengths of the unit cell for a rectangular box.
+
+    Notes
+    -----
+    * Cannot be defined alongside ``solvent_padding``,
+      ``number_of_solvent_molecules``, or ``box_vectors``.
+    """
+
+    @validator('box_vectors')
+    def supported_vectors(cls, v):
+        if v is not None:
+            if not _box_vectors_are_in_reduced_form(v):
+                errmsg = f"box_vectors: {v} are not in OpenMM reduced form"
+                raise ValueError(errmsg)
+        return v
 
     @validator('solvent_padding')
     def is_positive_distance(cls, v):
         # these are time units, not simulation steps
+        if v is None:
+            return v
+
         if not v.is_compatible_with(unit.nanometer):
             raise ValueError("solvent_padding must be in distance units "
                              "(i.e. nanometers)")
         if v < 0:
             errmsg = "solvent_padding must be a positive value"
             raise ValueError(errmsg)
+
+        return v
+
+    @validator('number_of_solvent_molecules')
+    def positive_solvent_number(cls, v):
+        if v is None:
+            return v
+
+        if v <= 0:
+            errmsg = f"number_of_solvent molecules: {v} must be positive"
+            raise ValueError(errmsg)
+
+        return v
+
+    @validator('box_size')
+    def box_size_properties(cls, v):
+        if v is None:
+            return v
+
+        if v.shape != (3, ):
+            errmsg = (f"box_size must be a 1-D array of length 3 "
+                      f"got {v} with shape {v.shape}")
+            raise ValueError(errmsg)
+
         return v
 
 
@@ -115,23 +258,27 @@ class OpenFFPartialChargeSettings(BasePartialChargeSettings):
     The OpenFF toolkit registry backend to use for partial charge generation.
 
 
-    OFF backend selection options
-    -----------------------------
+    OpenFF backend selection options
+    --------------------------------
 
-    The following are set depending on the option chosen:
-    * ``ambertools``: this will limit partial charge generation to using
-      a mixture of AmberTools and RDKit.
-    * ``openeye``: this will limit partial charge generation to using
-      the OpenEye toolkit. This cannot be used with ``espaloma`` as the
-      ``partial_charge_method``
-    * ``rdkit``: this will limit partial charge generation to using
-      the RDKit toolkit. Note that this alone cannot be used for conventionla
-      am1bcc partial charge generation, but is usually used in combination with
+    ``ambertools``:
+      This limits partial charge generation to using a mixture of AmberTools
+      and RDKit.
+
+    ``openeye``:
+      This limits partial charge generation to using the OpenEye toolkit.
+      This cannot be used with ``espaloma`` as the ``partial_charge_method``
+
+    ``rdkit``:
+      This limits partial charge generation to using the RDKit toolkit.
+      Note that this alone cannot be used for conventional am1bcc partial
+      charge generation, but is usually used in combination with
       the ``nagl`` or ``espaloma`` ``partial_charge_method`` selections.
+
     """
     number_of_conformers: Optional[int] = None
     """
-    Number of conformers to generate as part of the partial charge assignement.
+    Number of conformers to generate as part of the partial charge assignment.
 
     If ``None`` (default), the existing conformer of the input
     SmallMoleculeComponent will be used.
@@ -161,13 +308,20 @@ class OpenMMEngineSettings(SettingsBaseModel):
 
     compute_platform: Optional[str] = None
     """
-    OpenMM compute platform to perform MD integration with. If None, will
-    choose fastest available platform. Default None.
+    OpenMM compute platform to perform MD integration with. If ``None``, will
+    choose fastest available platform. Default ``None``.
     """
 
 
 class IntegratorSettings(SettingsBaseModel):
-    """Settings for the LangevinSplittingDynamicsMove integrator"""
+    """Settings for the LangevinDynamicsMove integrator
+
+    Note
+    ----
+    For some Protocols, an MC "move" (e.g. replica exchange swap) is applied
+    at a given frequency. In most Protocols the move frequency is defined in
+    `MultiStateSimulationSettings.time_per_iteration`.
+    """
 
     class Config:
         arbitrary_types_allowed = True
@@ -178,8 +332,8 @@ class IntegratorSettings(SettingsBaseModel):
     """Collision frequency. Default 1.0 / unit.pisecond."""
     reassign_velocities = False
     """
-    If True, velocities are reassigned from the Maxwell-Boltzmann
-    distribution at the beginning of move. Default False.
+    If ``True``, velocities are reassigned from the Maxwell-Boltzmann
+    distribution at the beginning of each MC move. Default ``False``.
     """
     n_restart_attempts = 20
     """
@@ -196,7 +350,7 @@ class IntegratorSettings(SettingsBaseModel):
     """
     remove_com: bool = False
     """
-    Whether or not to remove the center of mass motion. Default False.
+    Whether or not to remove the center of mass motion. Default ``False``.
     """
 
     @validator('langevin_collision_rate', 'n_restart_attempts')
@@ -245,7 +399,7 @@ class OutputSettings(SettingsBaseModel):
     Selection string for which part of the system to write coordinates for.
     Default 'not water'.
     """
-    checkpoint_interval: FloatQuantity['picosecond'] = 1 * unit.picosecond
+    checkpoint_interval: FloatQuantity['picosecond'] = 250 * unit.picosecond
     """
     Frequency to write the checkpoint file. Default 1 * unit.picosecond.
     """
@@ -298,18 +452,13 @@ class SimulationSettings(SettingsBaseModel):
     """Number of minimization steps to perform. Default 5000."""
     equilibration_length: FloatQuantity['nanosecond']
     """
-    Length of the equilibration phase in units of time. The total number of
-    steps from this equilibration length
-    (i.e. ``equilibration_length`` / :class:`IntegratorSettings.timestep`)
-    must be a multiple of the value defined for
-    :class:`AlchemicalSamplerSettings.steps_per_iteration`.
+    Length of the equilibration phase in units of time.
+    Must be divisible by the :class:`IntegratorSettings.timestep`.
     """
     production_length: FloatQuantity['nanosecond']
     """
-    Length of the production phase in units of time. The total number of
-    steps from this production length (i.e.
-    ``production_length`` / :class:`IntegratorSettings.timestep`) must be
-    a multiple of the value defined for :class:`IntegratorSettings.nsteps`.
+    Length of the production phase in units of time.
+    Must be divisible by the :class:`IntegratorSettings.timestep`.
     """
 
     @validator('equilibration_length', 'production_length')
@@ -340,10 +489,10 @@ class MultiStateSimulationSettings(SimulationSettings):
     ----
     * It'd be great if we could pass in the sampler object rather than using
       strings to define which one we want.
-    * Make n_replicas optional such that: If `None` or greater than the number
-      of lambda windows set in :class:`AlchemicalSettings`, this will default
-      to the number of lambda windows. If less than the number of lambda
-      windows, the replica lambda states will be picked at equidistant
+    * Make n_replicas optional such that: If ``None`` or greater than the
+      number of lambda windows set in :class:`LambdaSettings`, this will
+      default to the number of lambda windows. If less than the number of
+      lambda windows, the replica lambda states will be picked at equidistant
       intervals along the lambda schedule.
     """
 
