@@ -15,7 +15,7 @@ from openff.toolkit import Molecule as OFFMol
 from openff.toolkit.utils.toolkits import RDKitToolkitWrapper
 from openff.toolkit.utils.toolkit_registry import ToolkitRegistry
 from openff.units import unit
-from openff.units.openmm import ensure_quantity
+from openff.units.openmm import ensure_quantity, from_openmm
 from gufe.settings import OpenMMSystemGeneratorFFSettings, ThermoSettings
 import openfe
 from openfe.protocols.openmm_utils import (
@@ -30,6 +30,47 @@ from openfe.protocols.openmm_rfe.equil_rfe_settings import (
     OpenMMSolvationSettings,
 )
 from unittest import mock
+
+
+@pytest.mark.parametrize('padding, number_solv, box_vectors, box_size', [
+    [1.2 * unit.nanometer, 20, 20 * np.identity(3) * unit.angstrom,
+     [2, 2, 2] * unit.angstrom],
+    [1.2 * unit.nanometer, None, None, [2, 2, 2] * unit.angstrom],
+    [1.2 * unit.nanometer, None, 20 * np.identity(3) * unit.angstrom, None],
+    [1.2 * unit.nanometer, 20, None, None],
+])
+def test_validate_ommsolvation_settings_unique_settings(
+    padding, number_solv, box_vectors, box_size
+):
+    settings = OpenMMSolvationSettings(
+        solvent_padding=padding,
+        number_of_solvent_molecules=number_solv,
+        box_vectors=box_vectors,
+        box_size=box_size,
+    )
+
+    errmsg = "Only one of solvent_padding, number_of_solvent_molecules,"
+    with pytest.raises(ValueError, match=errmsg):
+        settings_validation.validate_openmm_solvation_settings(settings)
+
+
+@pytest.mark.parametrize('box_vectors, box_size', [
+    [20 * np.identity(3) * unit.angstrom, None],
+    [None, [2, 2, 2] * unit.angstrom],
+])
+def test_validate_ommsolvation_settings_shape_conflicts(
+    box_vectors, box_size,
+):
+    settings = OpenMMSolvationSettings(
+        solvent_padding=None,
+        box_vectors=box_vectors,
+        box_size=box_size,
+        box_shape='cube',
+    )
+
+    errmsg = "box_shape cannot be defined alongside either box_size"
+    with pytest.raises(ValueError, match=errmsg):
+        settings_validation.validate_openmm_solvation_settings(settings)
 
 
 def test_validate_timestep():
@@ -202,14 +243,14 @@ class TestFEAnalysis:
             yield r
         finally:
             r.close()
-    
+
     @pytest.fixture()
     def analyzer(self, reporter):
         return multistate_analysis.MultistateEquilFEAnalysis(
             reporter, sampling_method='repex',
             result_units=unit.kilocalorie_per_mole,
         )
-    
+
     def test_free_energies(self, analyzer):
         ret_dict = analyzer.unit_results_dict
         assert len(ret_dict.items()) == 7
@@ -256,7 +297,7 @@ class TestFEAnalysis:
             assert Path('replica_state_timeseries.png').is_file()
 
     def test_plot_convergence_bad_units(self, analyzer):
-        
+
         with pytest.raises(ValueError, match='Unknown plotting units'):
             openfe.analysis.plotting.plot_convergence(
                 analyzer.forward_and_reverse_free_energies,
@@ -356,18 +397,31 @@ class TestSystemCreation:
         assert_equal(comp_resids[openfe.SolventComponent()],
                      np.linspace(165, len(resids)-1, len(resids)-165))
 
-    def test_get_omm_modeller_ligand_no_neutralize(self, get_settings):
+    @pytest.fixture(scope='module')
+    def ligand_mol_and_generator(self, get_settings):
+        # Create offmol
+        offmol = OFFMol.from_smiles('[O-]C=O')
+        offmol.generate_conformers()
+        offmol.assign_partial_charges(partial_charge_method='am1bcc')
+        smc = openfe.SmallMoleculeComponent.from_openff(offmol)
+
         ffsets, intsets, thermosets = get_settings
         generator = system_creation.get_system_generator(
             ffsets, thermosets, intsets, None, True
         )
 
-        offmol = OFFMol.from_smiles('[O-]C=O')
-        offmol.generate_conformers()
-        smc = openfe.SmallMoleculeComponent.from_openff(offmol)
-
+        # Register offmol in generator
         generator.create_system(offmol.to_topology().to_openmm(),
                                 molecules=[offmol])
+
+        return (offmol, smc, generator)
+
+    def test_get_omm_modeller_ligand_no_neutralize(
+        self, ligand_mol_and_generator
+    ):
+
+        offmol, smc, generator = ligand_mol_and_generator
+
         model, comp_resids = system_creation.get_omm_modeller(
             None,
             openfe.SolventComponent(neutralize=False),
@@ -395,6 +449,90 @@ class TestSystemCreation:
 
         assert pytest.approx(charge.m) == -1.0
 
+    @pytest.mark.parametrize('n_expected, neutralize, shape',
+        [[400, False, 'cube'], [399, True, 'dodecahedron'],
+         [400, False, 'octahedron']]
+    )
+    def test_omm_modeller_ligand_n_solv(
+        self, ligand_mol_and_generator, n_expected, neutralize, shape
+    ):
+        offmol, smc, generator = ligand_mol_and_generator
+
+        solv_settings = OpenMMSolvationSettings(
+            solvent_padding=None,
+            number_of_solvent_molecules=400,
+            box_vectors=None,
+            box_size=None,
+            box_shape=shape
+        )
+
+        model, comp_resids = system_creation.get_omm_modeller(
+            None,
+            openfe.SolventComponent(
+                neutralize=neutralize,
+                ion_concentration = 0 * unit.molar
+            ),
+            {smc: offmol},
+            generator.forcefield,
+            solv_settings,
+        )
+
+        waters = [r for r in model.topology.residues() if r.name == 'HOH']
+        assert len(waters) == n_expected
+
+    def test_omm_modeller_box_size(self, ligand_mol_and_generator):
+        offmol, smc, generator = ligand_mol_and_generator
+
+        solv_settings = OpenMMSolvationSettings(
+            solvent_padding=None,
+            number_of_solvent_molecules=None,
+            box_vectors=None,
+            box_size=[2, 2, 2]*unit.nanometer,
+            box_shape=None
+        )
+
+        model, comp_resids = system_creation.get_omm_modeller(
+            None,
+            openfe.SolventComponent(),
+            {smc: offmol},
+            generator.forcefield,
+            solv_settings
+        )
+
+        vectors = model.topology.getPeriodicBoxVectors()
+
+        assert_allclose(
+            from_openmm(vectors),
+            [[2, 0, 0], [0, 2, 0], [0, 0, 2]] * unit.nanometer
+        )
+
+    def test_omm_modeller_box_vectors(self, ligand_mol_and_generator):
+        offmol, smc, generator = ligand_mol_and_generator
+
+        solv_settings = OpenMMSolvationSettings(
+            solvent_padding=None,
+            number_of_solvent_molecules=None,
+            box_vectors=[
+                [2, 0, 0], [0, 2, 0], [0, 0, 5]
+            ] * unit.nanometer,
+            box_size=None,
+            box_shape=None,
+        )
+
+        model, comp_resids = system_creation.get_omm_modeller(
+            None,
+            openfe.SolventComponent(),
+            {smc: offmol},
+            generator.forcefield,
+            solv_settings
+        )
+
+        vectors = model.topology.getPeriodicBoxVectors()
+
+        assert_allclose(
+            from_openmm(vectors),
+            [[2, 0, 0], [0, 2, 0], [0, 0, 5]] * unit.nanometer
+        )
 
 def test_convert_steps_per_iteration():
     sim = omm_settings.MultiStateSimulationSettings(
@@ -515,7 +653,7 @@ class TestOFFPartialCharge:
         ] * unit.elementary_charge
 
         uncharged_mol.partial_charges = copy.deepcopy(chg)
-    
+
         charge_generation.assign_offmol_partial_charges(
             uncharged_mol,
             overwrite=overwrite,
@@ -524,7 +662,7 @@ class TestOFFPartialCharge:
             generate_n_conformers=None,
             nagl_model=None,
         )
-    
+
         assert np.allclose(uncharged_mol.partial_charges, chg) != overwrite
 
     def test_unknown_method(self, uncharged_mol):
@@ -588,7 +726,7 @@ class TestOFFPartialCharge:
             )
 
     def test_too_many_requested_conformers(self, uncharged_mol):
-        
+
         with pytest.raises(ValueError, match="5 conformers were requested"):
             charge_generation.assign_offmol_partial_charges(
                 uncharged_mol,
@@ -651,7 +789,7 @@ class TestOFFPartialCharge:
 
     @pytest.mark.skipif(not HAS_NAGL, reason='NAGL is not available')
     def test_no_production_nagl(self, uncharged_mol):
-        
+
         with pytest.raises(ValueError, match='No production am1bcc NAGL'):
             charge_generation.assign_offmol_partial_charges(
                 uncharged_mol,
@@ -789,6 +927,8 @@ class TestOFFPartialCharge:
 
 @pytest.mark.slow
 @pytest.mark.download
+# Sometimes we get a DOI lookup error from duecredit
+@pytest.mark.flaky(reruns=3, only_rerun=ValueError, reruns_delay=10)
 def test_forward_backwards_failure(simulation_nc):
     rep = multistate.multistatereporter.MultiStateReporter(
         simulation_nc,
