@@ -7,6 +7,7 @@ from ...conftest import atom_mapping_basic_test_files, T4_protein_component
 from gufe import SolventComponent, AlchemicalNetwork
 from openfe.setup.alchemical_network_planner import RHFEAlchemicalNetworkPlanner, RBFEAlchemicalNetworkPlanner
 from .edge_types import r_complex_edge, r_solvent_edge, r_vacuum_edge
+import openfe
 
 
 def test_rhfe_alchemical_network_planner_init():
@@ -22,7 +23,15 @@ def test_rbfe_alchemical_network_planner_init():
 
 
 def test_rbfe_alchemical_network_planner_call(atom_mapping_basic_test_files, T4_protein_component):
-    alchem_planner = RBFEAlchemicalNetworkPlanner()
+    alchem_planner = RBFEAlchemicalNetworkPlanner(
+        # these aren't default settings, but the test ligands aren't aligned
+        mappers=[openfe.LomapAtomMapper(
+            time=20,
+            element_change=False,
+            max3d=1,
+            shift=True,
+        )]
+    )
     alchem_network = alchem_planner(
         ligands=atom_mapping_basic_test_files.values(),
         solvent=SolventComponent(),
@@ -34,14 +43,20 @@ def test_rbfe_alchemical_network_planner_call(atom_mapping_basic_test_files, T4_
     edges = alchem_network.edges
     assert len(edges) == 14 # we build 2envs*8ligands-2startLigands = 14 relative edges. 
 
-    print(edges)
     assert sum([r_complex_edge(e) for e in edges]) == 7 # half of the transformations should be complex (they always are)!
     assert sum([r_solvent_edge(e) for e in edges]) == 7 # half of the transformations should be solvent!
     assert sum([r_vacuum_edge(e) for e in edges]) == 0 # no vacuum here!
 
 
 def test_rhfe_alchemical_network_planner_call_multigraph(atom_mapping_basic_test_files):
-    alchem_planner = RHFEAlchemicalNetworkPlanner()
+    alchem_planner = RHFEAlchemicalNetworkPlanner(
+        mappers=[openfe.LomapAtomMapper(
+            time=20,
+            element_change=False,
+            max3d=1,
+            shift=True,
+        )]
+    )
 
     ligand_network = alchem_planner._construct_ligand_network(atom_mapping_basic_test_files.values())
     ligand_network_edges = list(ligand_network.edges)
@@ -59,7 +74,14 @@ def test_rhfe_alchemical_network_planner_call_multigraph(atom_mapping_basic_test
 
 
 def test_rhfe_alchemical_network_planner_call(atom_mapping_basic_test_files):
-    alchem_planner = RHFEAlchemicalNetworkPlanner()
+    alchem_planner = RHFEAlchemicalNetworkPlanner(
+        mappers=[openfe.LomapAtomMapper(
+            time=20,
+            element_change=False,
+            max3d=1,
+            shift=True,
+        )]
+    )
     alchem_network = alchem_planner(ligands=atom_mapping_basic_test_files.values(), solvent=SolventComponent())
     
     assert isinstance(alchem_network, AlchemicalNetwork)
