@@ -11,12 +11,26 @@ from openfe.setup.atom_mapping import LomapAtomMapper
 from .conftest import mol_from_smiles
 
 
-def test_simple(atom_mapping_basic_test_files):
+@pytest.fixture(scope='module')
+def lomap_old_mapper():
+    """
+    Old default options for the LomapAtomMapper.
+
+    This is necessary as the atom_mapping_basic_test_files
+    are not aligned and need shift & a large max3d value
+    """
+    return LomapAtomMapper(
+        time=20, threed=True, max3d=1000.0,
+        element_change=True, seed='', shift=True
+    )
+
+
+def test_simple(atom_mapping_basic_test_files, lomap_old_mapper):
     # basic sanity check on the LigandAtomMapper
     mol1 = atom_mapping_basic_test_files['methylcyclohexane']
     mol2 = atom_mapping_basic_test_files['toluene']
 
-    mapper = LomapAtomMapper()
+    mapper = lomap_old_mapper
 
     mapping_gen = mapper.suggest_mappings(mol1, mol2)
 
@@ -26,12 +40,12 @@ def test_simple(atom_mapping_basic_test_files):
     assert len(mapping.componentA_to_componentB) == 15
 
 
-def test_distances(atom_mapping_basic_test_files):
+def test_distances(atom_mapping_basic_test_files, lomap_old_mapper):
     # basic sanity check on the LigandAtomMapper
     mol1 = atom_mapping_basic_test_files['methylcyclohexane']
     mol2 = atom_mapping_basic_test_files['toluene']
 
-    mapper = LomapAtomMapper()
+    mapper = lomap_old_mapper
     mapping = next(mapper.suggest_mappings(mol1, mol2))
 
     dists = mapping.get_distances()
@@ -45,13 +59,13 @@ def test_distances(atom_mapping_basic_test_files):
     assert pytest.approx(dists[0], rel=1e-5) == 0.07249779
 
 
-def test_generator_length(atom_mapping_basic_test_files):
+def test_generator_length(atom_mapping_basic_test_files, lomap_old_mapper):
     # check that we get one mapping back from Lomap LigandAtomMapper then the
     # generator stops correctly
     mol1 = atom_mapping_basic_test_files['methylcyclohexane']
     mol2 = atom_mapping_basic_test_files['toluene']
 
-    mapper = LomapAtomMapper()
+    mapper = lomap_old_mapper
 
     mapping_gen = mapper.suggest_mappings(mol1, mol2)
 
@@ -60,12 +74,12 @@ def test_generator_length(atom_mapping_basic_test_files):
         next(mapping_gen)
 
 
-def test_bad_mapping(atom_mapping_basic_test_files):
+def test_bad_mapping(atom_mapping_basic_test_files, lomap_old_mapper):
     toluene = atom_mapping_basic_test_files['toluene']
     NigelTheNitrogen = SmallMoleculeComponent(mol_from_smiles('N'),
                                               name='Nigel')
 
-    mapper = LomapAtomMapper()
+    mapper = lomap_old_mapper
 
     mapping_gen = mapper.suggest_mappings(toluene, NigelTheNitrogen)
     with pytest.raises(StopIteration):
@@ -73,12 +87,14 @@ def test_bad_mapping(atom_mapping_basic_test_files):
 
 
 # TODO: Remvoe these test when element changes are allowed - START
-def test_simple_no_element_changes(atom_mapping_basic_test_files):
+def test_simple_no_element_changes(
+    atom_mapping_basic_test_files, lomap_old_mapper
+):
     # basic sanity check on the LigandAtomMapper
     mol1 = atom_mapping_basic_test_files['methylcyclohexane']
     mol2 = atom_mapping_basic_test_files['toluene']
 
-    mapper = LomapAtomMapper()
+    mapper = lomap_old_mapper
     mapper._no_element_changes = True
     mapping_gen = mapper.suggest_mappings(mol1, mol2)
 
@@ -88,12 +104,14 @@ def test_simple_no_element_changes(atom_mapping_basic_test_files):
     assert len(mapping.componentA_to_componentB) == 15
 
     
-def test_bas_mapping_no_element_changes(atom_mapping_basic_test_files):
+def test_bas_mapping_no_element_changes(
+    atom_mapping_basic_test_files, lomap_old_mapper
+):
     toluene = atom_mapping_basic_test_files['toluene']
     NigelTheNitrogen = SmallMoleculeComponent(mol_from_smiles('N'),
                                               name='Nigel')
 
-    mapper = LomapAtomMapper()
+    mapper = lomap_old_mapper
     mapper._no_element_changes = True
     mapping_gen = mapper.suggest_mappings(toluene, NigelTheNitrogen)
     with pytest.raises(StopIteration):
