@@ -189,7 +189,10 @@ def test_heterocycle_score(base, other, name, hit):
     m1 = openfe.SmallMoleculeComponent.from_rdkit(r1)
     m2 = openfe.SmallMoleculeComponent.from_rdkit(r2)
 
-    mapper = openfe.setup.atom_mapping.LomapAtomMapper(threed=False)
+    mapper = openfe.setup.atom_mapping.LomapAtomMapper(
+        time=20, threed=False, max3d=1000.0,
+        element_change=True, seed='', shift=True,
+    )
     mapping = next(mapper.suggest_mappings(m1, m2))
     score = lomap_scorers.heterocycles_score(mapping)
 
@@ -223,7 +226,10 @@ def test_lomap_individual_scores(params,
                                       molB.to_rdkit()), scorename)()
 
     # longer way
-    mapper = openfe.setup.atom_mapping.LomapAtomMapper(threed=False)
+    mapper = openfe.setup.atom_mapping.LomapAtomMapper(
+        time=20, threed=False, max3d=1000.0,
+        element_change=True, seed='', shift=True,
+    )
     mapping = next(mapper.suggest_mappings(molA, molB))
     openfe_version = getattr(lomap_scorers, SCORE_NAMES[scorename])(mapping)
 
@@ -248,7 +254,10 @@ def test_lomap_regression(lomap_basic_test_files_dir,  # in a dir for lomap
         nm = dbmols[i].getName()
         smallmols.append(atom_mapping_basic_test_files[nm[:-5]])  # - ".mol2"
 
-    mapper = openfe.setup.atom_mapping.LomapAtomMapper(threed=False)
+    mapper = openfe.setup.atom_mapping.LomapAtomMapper(
+        time=20, threed=False, max3d=1000.0,
+        element_change=True, seed='', shift=True,
+    )
     scorer = lomap_scorers.default_lomap_score
     scores = np.zeros_like(matrix)
     for i, j in itertools.combinations(range(matrix.shape[0]), 2):
