@@ -2,14 +2,12 @@
 # For details, see https://github.com/OpenFreeEnergy/openfe
 import pytest
 from openfe.setup.atom_mapping import PersesAtomMapper, LigandAtomMapping
+from openff.units import unit
 
 pytest.importorskip('perses')
 pytest.importorskip('openeye')
 
-USING_NEW_OFF = True  # by default we are now
 
-
-@pytest.mark.xfail(USING_NEW_OFF, reason="Perses #1108")
 def test_simple(atom_mapping_basic_test_files):
     # basic sanity check on the LigandAtomMapper
     mol1 = atom_mapping_basic_test_files['methylcyclohexane']
@@ -25,7 +23,6 @@ def test_simple(atom_mapping_basic_test_files):
     assert len(mapping.componentA_to_componentB) == 4
 
 
-@pytest.mark.xfail(USING_NEW_OFF, reason="Perses #1108")
 def test_generator_length(atom_mapping_basic_test_files):
     # check that we get one mapping back from Lomap LigandAtomMapper then the
     # generator stops correctly
@@ -41,7 +38,6 @@ def test_generator_length(atom_mapping_basic_test_files):
         next(mapping_gen)
 
 
-@pytest.mark.xfail(USING_NEW_OFF, reason="Perses #1108")
 def test_empty_atommappings(mol_pair_to_shock_perses_mapper):
     mol1, mol2 = mol_pair_to_shock_perses_mapper
     mapper = PersesAtomMapper()
@@ -53,3 +49,14 @@ def test_empty_atommappings(mol_pair_to_shock_perses_mapper):
 
     with pytest.raises(StopIteration):
         next(mapping_gen)
+
+
+def test_dict_round_trip():
+    # use some none defaults
+    mapper1 = PersesAtomMapper(
+        allow_ring_breaking=False,
+        preserve_chirality=False,
+        coordinate_tolerance=0.01 * unit.nanometer
+    )
+    mapper2 = PersesAtomMapper.from_dict(mapper1.to_dict())
+    assert mapper2.to_dict() == mapper1.to_dict()
