@@ -31,24 +31,24 @@ def tyk2_protein():
 
 @pytest.fixture
 def expected_transformations():
-    return ['easy_rbfe_lig_ejm_31_complex_lig_ejm_42_complex.json',
-            'easy_rbfe_lig_ejm_31_solvent_lig_ejm_50_solvent.json',
-            'easy_rbfe_lig_ejm_31_complex_lig_ejm_46_complex.json',
-            'easy_rbfe_lig_ejm_42_complex_lig_ejm_43_complex.json',
+    return ['easy_rbfe_lig_ejm_31_complex_lig_ejm_46_complex.json',
             'easy_rbfe_lig_ejm_31_complex_lig_ejm_47_complex.json',
-            'easy_rbfe_lig_ejm_42_solvent_lig_ejm_43_solvent.json',
             'easy_rbfe_lig_ejm_31_complex_lig_ejm_48_complex.json',
-            'easy_rbfe_lig_ejm_46_complex_lig_jmc_23_complex.json',
             'easy_rbfe_lig_ejm_31_complex_lig_ejm_50_complex.json',
-            'easy_rbfe_lig_ejm_46_complex_lig_jmc_27_complex.json',
-            'easy_rbfe_lig_ejm_31_solvent_lig_ejm_42_solvent.json',
-            'easy_rbfe_lig_ejm_46_complex_lig_jmc_28_complex.json',
             'easy_rbfe_lig_ejm_31_solvent_lig_ejm_46_solvent.json',
-            'easy_rbfe_lig_ejm_46_solvent_lig_jmc_23_solvent.json',
             'easy_rbfe_lig_ejm_31_solvent_lig_ejm_47_solvent.json',
-            'easy_rbfe_lig_ejm_46_solvent_lig_jmc_27_solvent.json',
             'easy_rbfe_lig_ejm_31_solvent_lig_ejm_48_solvent.json',
-            'easy_rbfe_lig_ejm_46_solvent_lig_jmc_28_solvent.json']
+            'easy_rbfe_lig_ejm_31_solvent_lig_ejm_50_solvent.json',
+            'easy_rbfe_lig_ejm_42_complex_lig_ejm_43_complex.json',
+            'easy_rbfe_lig_ejm_42_complex_lig_ejm_50_complex.json',
+            'easy_rbfe_lig_ejm_42_solvent_lig_ejm_43_solvent.json',
+            'easy_rbfe_lig_ejm_42_solvent_lig_ejm_50_solvent.json',
+            'easy_rbfe_lig_ejm_46_complex_lig_jmc_27_complex.json',
+            'easy_rbfe_lig_ejm_46_solvent_lig_jmc_27_solvent.json',
+            'easy_rbfe_lig_jmc_23_complex_lig_jmc_27_complex.json',
+            'easy_rbfe_lig_jmc_23_solvent_lig_jmc_27_solvent.json',
+            'easy_rbfe_lig_jmc_27_complex_lig_jmc_28_complex.json',
+            'easy_rbfe_lig_jmc_27_solvent_lig_jmc_28_solvent.json']
 
 
 def test_plan_tyk2(tyk2_ligands, tyk2_protein, expected_transformations):
@@ -88,15 +88,15 @@ def mock_execute(expected_transformations):
 def ref_gather():
     return """\
 ligand_i\tligand_j\tDDG(i->j) (kcal/mol)\tuncertainty (kcal/mol)
-lig_ejm_31\tlig_ejm_42\t0.0\t0.0
 lig_ejm_31\tlig_ejm_46\t0.0\t0.0
 lig_ejm_31\tlig_ejm_47\t0.0\t0.0
 lig_ejm_31\tlig_ejm_48\t0.0\t0.0
 lig_ejm_31\tlig_ejm_50\t0.0\t0.0
 lig_ejm_42\tlig_ejm_43\t0.0\t0.0
-lig_ejm_46\tlig_jmc_23\t0.0\t0.0
+lig_ejm_42\tlig_ejm_50\t0.0\t0.0
 lig_ejm_46\tlig_jmc_27\t0.0\t0.0
-lig_ejm_46\tlig_jmc_28\t0.0\t0.0
+lig_jmc_23\tlig_jmc_27\t0.0\t0.0
+lig_jmc_27\tlig_jmc_28\t0.0\t0.0
 """
 
 
